refactor(harness): move embeddings OpenAI provider out of inline mod
embeddings/mod.rs embedded a full HTTP provider (OpenAiEmbeddingModel)
as an inline `mod openai { ... }` block instead of its own file. Moved
to openai.rs, declared via `mod openai;`, mirroring the convention
used by harness::providers::openai. No behavior change.
